requirements.txt
Django>=5.0,<5.1
djangorestframework
django-redis
gunicorn
psycopg2-binary>=2.9.9
valkey>=6.0.0
djangorestframework>=3.15.1
django-cors-headers>=4.3.1
python-dotenv>=1.0.0
dj-database-url>=2.1.0
whitenoise>=6.6.0
PyJWT>=2.8.0
cryptography>=42.0.0
nats-py>=2.7.2
